relocate and update to comments
