+++ manage-todos-with-realtime-agent/todo_utils.py
@@ -1,6 +1,8 @@
 import json
 from typing import Annotated
+from logging import getLogger
 
+logger = getLogger(__name__)
 
 def format_todo_str(data):
     todo_str = "Current todos:\n"
@@ -31,7 +33,6 @@ def add_todo(
     # if exist, load the data
     # if not exist, create empty list
     data = get_data()
-
     data.append(
         {
             "id": id,
@@ -39,9 +40,7 @@ def add_todo(
             "status": status,
         }
     )
-
     save_data(data)
-
     return f"Todo with id {id} added.\n" + format_todo_str(data)
 
 
@@ -49,7 +48,7 @@ def modify_todo(
     id: Annotated[int, "task id"],
     task: Annotated[str, "task description"],
     status: Annotated[str, "task status, either done or open"],
-):
+) -> str:
     data = get_data()
     is_todo_found = False
     for todo in data:
@@ -68,7 +67,12 @@ def modify_todo(
     return f"Todo with id {id} updated.\n" + format_todo_str(data)
 
 
-def delete_todo(id: Annotated[int, "task id"]):
+def delete_todo(
+    id: Annotated[int, "task id"],
+    task: Annotated[str, "task description"],
+    status: Annotated[str, "task status, either done or open"],
+) -> str:
+    logger.info(f"Deleting todo with id {id}, task {task}")
     data = get_data()
 
     is_todo_found = False
@@ -79,7 +83,14 @@ def delete_todo(id: Annotated[int, "task id"]):
         else:
             new_data.append(todo)
 
+    data = new_data
     save_data(data)
     if is_todo_found:
-        return f"Todo with id {id} deleted.\n" + format_todo_str(data)
-    return f"Todo with id {id} not found. Nothing changed.\n" + format_todo_str(data)
+        msg = f"Todo with id {id} deleted.\n" + format_todo_str(data)
+    else:
+        msg = f"Todo with id {id} not found. Nothing changed.\n" + format_todo_str(
+            data
+        )
+    logger.info(f"Msg from delete: {msg}")
+    return msg
+
